// src/types/CategoryFiltered.ts
export enum CategoryFiltered {
  ALL = 'All',
  SIDE = 'Side',
  MEAT = 'Meat',
  VEGETARIAN = 'Vegetarian',
  DESSERT = 'Dessert',
  SEAFOOD = 'Seafood',
  PASTA = 'Pasta',
}

export const CATEGORY_OPTIONS = {
  All: 'All',
  Side : 'Side',
  Meat : 'Meat',
  Vegetarian : 'Vegetarian',
  Dessert : 'Dessert',
}